% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{enable_auto_alt_text}
\alias{enable_auto_alt_text}
\title{Add alt text to Quarto figures}
\usage{
enable_auto_alt_text(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=generate_alt_text]{generate_alt_text()}}.}
}
\value{
NULL, invisibly
}
\description{
Call this function once at the top of your Quarto document
to include an alternative text for all of your ggplot2 charts.
}
