% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{generate_alt_text}
\alias{generate_alt_text}
\title{Generate alternative text from a ggplot2 object}
\usage{
generate_alt_text(p, from = c("default", "auto", "origin"))
}
\arguments{
\item{p}{A \code{ggplot} object.}

\item{from}{Method used to select alt text. Since ggplot2 objects can
include built-in alt text (via \code{labs(alt = "Here goes the text")}),
choose whether to use that text or the one generated by \code{ggalttext}.
- \code{"default"}: always generate alt text with \code{ggalttext}.
- \code{"auto"}: use existing alt text when it is non-empty; otherwise generate
alt text with \code{ggalttext}.
- \code{"origin"}: always return \code{ggplot2::get_alt_text(p)} as-is. In that case,
you may not need \code{ggalttext}.}
}
\value{
A string
}
\description{
\code{generate_alt_text()} takes any ggplot2 object and generates
a string that describes what contains the chart.

The description isn't meant to exactly describe the chart,
nor how it was made. Instead, it tries to be short and
gives an overview of what's inside the chart such as:
\itemize{
\item the kind of chart(s)
\item the number of chart(s) for facets
\item the title, subtitle and caption
}

Learn more: https://www.section508.gov/create/alternative-text/
}
