% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alt_text.R
\name{generate_alt_text}
\alias{generate_alt_text}
\title{Generate Alt Text}
\usage{
generate_alt_text(g)
}
\arguments{
\item{g}{A ggplot object.}
}
\value{
A character string containing alternative text describing the input ggplot object.
}
\description{
This function generates alternative text for ggplot objects.
}
\examples{
library(ggplot2)
g1 <- ggplot(data = mtcars,
             mapping = aes(x = mpg,
                           y = disp)) +
  geom_point()
generate_alt_text(g1)
}
