% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{generate_alt_text}
\alias{generate_alt_text}
\title{Generate alternative text from a ggplot2 object}
\usage{
generate_alt_text(p)
}
\arguments{
\item{p}{A ggplot2 chart}
}
\value{
A string
}
\description{
\code{generate_alt_text()} takes any ggplot2 object and generate
a string that describes what contains the chart.

The description isn't meant to exactly describe the chart,
nor how it was made. Instead, it tries to be short and
gives an overview of what's inside the chart such as:
\itemize{
\item the kind of chart(s)
\item the number of chart(s) for facets
\item the title, subtitle and caption
}

Learn more: https://www.section508.gov/create/alternative-text/
}
