% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{generate_alt_text}
\alias{generate_alt_text}
\title{Generate alternative text from a ggplot2 object}
\usage{
generate_alt_text(
  p,
  ...,
  lang = c("en", "fr", "de"),
  include_title = TRUE,
  max_chars = 125,
  error_on_excess = FALSE,
  source = c("default", "auto", "origin")
)
}
\arguments{
\item{p}{A \code{ggplot} object.}

\item{...}{Must be empty. This argument requires all arguments after \code{p} to
be named.}

\item{lang}{Language used for generated text. Defaults to \code{"en"} (English).
Supported values are \code{"en"}, \code{"fr"} (French), and \code{"de"} (German).}

\item{include_title}{Whether to include the title of the plot, if any, in the
alternative text. This is mostly useful to reduce the length of the text.}

\item{max_chars}{The maximum number of characters allowed (default to 125 as
recommended by most references). If above, it will raise a warning. If
\code{error_on_excess = TRUE}, then it will raise an error.}

\item{error_on_excess}{Whether to raise an error when the number of
characters of the alt text is above \code{max_chars}. If \code{FALSE}, it will
raise a warning.}

\item{source}{Method used to select alt text. Since ggplot2 objects can
include built-in alt text (via \code{labs(alt = "Here goes the text")}),
choose whether to use that text or the one generated by \code{ggalttext}.
- \code{"default"}: always generate alt text with \code{ggalttext}.
- \code{"auto"}: use existing alt text when it is non-empty; otherwise generate
alt text with \code{ggalttext}.
- \code{"origin"}: always return \code{ggplot2::get_alt_text(p)} as-is.}
}
\value{
A string
}
\description{
\code{generate_alt_text()} takes any ggplot2 object and generates
a string that describes what contains the chart.

The description isn't meant to exactly describe the chart,
nor how it was made. Instead, it tries to be short and
gives an overview of what's inside the chart such as:
\itemize{
\item the kind of chart(s)
\item the number of chart(s) for facets
\item the title
\item ...
}

The text can be generated in: English (default), French or
German.

Learn more: https://www.section508.gov/create/alternative-text/
}
